move create response data to new file
